fix list and prune command

+++ packages/core/src/bin/commands/prune.ts
@@ -1,6 +1,7 @@
 import { createBuild } from "@/build/index.js";
 import {
   type PonderApp,
+  VIEWS,
   createDatabase,
   getPonderMetaTable,
 } from "@/database/index.js";
@@ -77,6 +78,11 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
     )
     .groupBy(TABLES.table_schema);
 
+  const ponderViewSchemas = await database.qb.drizzle
+    .select({ schema: VIEWS.table_schema })
+    .from(VIEWS)
+    .where(eq(VIEWS.table_name, "_ponder_meta"));
+
   const queries = ponderSchemas.map((row) =>
     database.qb.drizzle
       .select({
@@ -106,6 +112,7 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
   }
 
   const tablesToDrop: string[] = [];
+  const viewsToDrop: string[] = [];
   const schemasToDrop: string[] = [];
   const functionsToDrop: string[] = [];
 
@@ -118,28 +125,51 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
       continue;
     }
 
-    for (const table of value.table_names) {
-      tablesToDrop.push(`"${schema}"."${table}"`);
-      tablesToDrop.push(`"${schema}"."${sqlToReorgTableName(table)}"`);
-      functionsToDrop.push(`"${schema}"."operation_reorg__${table}"`);
-    }
-    tablesToDrop.push(`"${schema}"."_ponder_meta"`);
-    if (value.version === "2") {
-      tablesToDrop.push(`"${schema}"."_ponder_checkpoint"`);
+    if (ponderViewSchemas.some((vs) => vs.schema === schema)) {
+      for (const table of value.table_names) {
+        viewsToDrop.push(`"${schema}"."${table}"`);
+      }
+      viewsToDrop.push(`"${schema}"."_ponder_meta"`);
+      if (value.version === "2") {
+        viewsToDrop.push(`"${schema}"."_ponder_checkpoint"`);
+      } else {
+        viewsToDrop.push(`"${schema}"."_ponder_status"`);
+      }
+
+      const tableCount = ponderSchemas.find(
+        (s) => s.schema === schema,
+      )!.tableCount;
+
+      if (schema !== "public" && tableCount <= 2 + value.table_names.length) {
+        schemasToDrop.push(schema);
+      }
     } else {
-      tablesToDrop.push(`"${schema}"."_ponder_status"`);
-    }
+      for (const table of value.table_names) {
+        tablesToDrop.push(`"${schema}"."${table}"`);
+        tablesToDrop.push(`"${schema}"."${sqlToReorgTableName(table)}"`);
+        functionsToDrop.push(`"${schema}"."operation_reorg__${table}"`);
+      }
+      tablesToDrop.push(`"${schema}"."_ponder_meta"`);
+      if (value.version === "2") {
+        tablesToDrop.push(`"${schema}"."_ponder_checkpoint"`);
+      } else {
+        tablesToDrop.push(`"${schema}"."_ponder_status"`);
+      }
 
-    const tableCount = ponderSchemas.find(
-      (s) => s.schema === schema,
-    )!.tableCount;
+      const tableCount = ponderSchemas.find(
+        (s) => s.schema === schema,
+      )!.tableCount;
 
-    if (schema !== "public" && tableCount <= 2 + value.table_names.length * 2) {
-      schemasToDrop.push(schema);
+      if (
+        schema !== "public" &&
+        tableCount <= 2 + value.table_names.length * 2
+      ) {
+        schemasToDrop.push(schema);
+      }
     }
   }
 
-  if (tablesToDrop.length === 0) {
+  if (tablesToDrop.length === 0 && viewsToDrop.length === 0) {
     logger.warn({
       service: "prune",
       msg: "No inactive Ponder apps found in this database.",
@@ -148,23 +178,38 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
     return;
   }
 
-  await database.qb.drizzle.execute(
-    sql.raw(`DROP TABLE IF EXISTS ${tablesToDrop.join(", ")} CASCADE`),
-  );
+  if (tablesToDrop.length > 0) {
+    await database.qb.drizzle.execute(
+      sql.raw(`DROP TABLE IF EXISTS ${tablesToDrop.join(", ")} CASCADE`),
+    );
 
-  logger.warn({
-    service: "prune",
-    msg: `Dropped ${tablesToDrop.length} tables`,
-  });
+    logger.warn({
+      service: "prune",
+      msg: `Dropped ${tablesToDrop.length} tables`,
+    });
+  }
 
-  await database.qb.drizzle.execute(
-    sql.raw(`DROP FUNCTION IF EXISTS ${functionsToDrop.join(", ")} CASCADE`),
-  );
+  if (viewsToDrop.length > 0) {
+    await database.qb.drizzle.execute(
+      sql.raw(`DROP VIEW IF EXISTS ${viewsToDrop.join(", ")} CASCADE`),
+    );
 
-  logger.warn({
-    service: "prune",
-    msg: `Dropped ${functionsToDrop.length} functions`,
-  });
+    logger.warn({
+      service: "prune",
+      msg: `Dropped ${viewsToDrop.length} views`,
+    });
+  }
+
+  if (functionsToDrop.length > 0) {
+    await database.qb.drizzle.execute(
+      sql.raw(`DROP FUNCTION IF EXISTS ${functionsToDrop.join(", ")} CASCADE`),
+    );
+
+    logger.warn({
+      service: "prune",
+      msg: `Dropped ${functionsToDrop.length} functions`,
+    });
+  }
 
   if (schemasToDrop.length > 0) {
     await database.qb.drizzle.execute(
